api-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
pyjwt==2.8.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
prometheus-client==0.19.0
